requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
faiss-cpu
scikit-learn
google-generativeai
pytz
deep_translator
gTTS
speechrecognition
pydub
tavily-python
python-dotenv
pandas
openpyxl
PyPDF2
sentence-transformers
